move authorship inherents into primitives
