+++ Cargo.toml
@@ -89,6 +89,7 @@ members = [
 	"paint/utility",
 	"primitives/application-crypto",
 	"primitives/authority-discovery",
+	"primitives/authorship",
 	"primitives/block-builder/runtime-api",
 	"primitives/consensus/aura",
 	"primitives/consensus/babe",
@@ -122,6 +123,7 @@ members = [
 	"primitives/sr-std",
 	"primitives/sr-version",
 	"primitives/state-machine",
+	"primitives/timestamp",
 	"primitives/transaction-pool/runtime-api",
 	"primitives/trie",
 	"primitives/wasm-interface",
