+++ paint/authorship/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2018"
 primitives = { package = "substrate-primitives",  path = "../../primitives/core", default-features = false }
 codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
 inherents = { package = "substrate-inherents", path = "../../primitives/inherents", default-features = false }
+sp-authorship = { path = "../../primitives/authorship", default-features = false }
 rstd = { package = "sr-std", path = "../../primitives/sr-std", default-features = false }
 sr-primitives = { path = "../../primitives/sr-primitives", default-features = false }
 support = { package = "paint-support", path = "../support", default-features = false }
@@ -27,4 +28,5 @@ std = [
 	"support/std",
 	"system/std",
 	"runtime-io/std",
+	"sp-authorship/std",
 ]
