+++ primitives/authorship/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "sp-authorship"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+description = "Authorship primitives"
+edition = "2018"
+
+[dependencies]
+sp-inherents = { package = "substrate-inherents", path = "../inherents", default-features = false }
+rstd = { package = "sr-std", path = "../sr-std", default-features = false }
+codec = { package = "parity-scale-codec", version = "1.0.0", default-features = false, features = ["derive"] }
+
+[features]
+default = [ "std" ]
+std = [
+    "codec/std",
+    "rstd/std",
+    "sp-inherents/std",
+]
